+++ app/src/main/res/layout/activity_sort.xml
@@ -52,7 +52,7 @@
             <androidx.recyclerview.widget.RecyclerView
                 android:id="@+id/sorting"
                 android:layout_width="match_parent"
-                android:layout_height="100dp"
+                android:layout_height="200dp"
                 />
 
             <TextView
