--- app/src/main/res/layout/delafter.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-    <!-- list_item_layout.xml -->
-    <TextView xmlns:android="http://schemas.android.com/apk/res/android"
-        android:id="@+id/cityNameTextView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:textSize="16sp"
-        android:padding="8dp" />
