+++ app/src/main/res/layout/sort_content.xml
@@ -1,0 +1,39 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:orientation="horizontal"
+    android:padding="16dp">
+
+    <!-- Checkbox on the left side -->
+    <CheckBox
+        android:id="@+id/checkBox"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center_vertical"
+        android:layout_marginEnd="8dp" />
+
+    <!-- Text view for the sort name -->
+    <TextView
+        android:id="@+id/sortTextView"
+        android:layout_width="0dp"
+        android:layout_height="match_parent"
+        android:layout_weight="1"
+        android:textSize="16sp"
+        android:gravity="center"
+        android:text="@string/sort_method_name"
+        android:textColor="@android:color/black" />
+
+    <!-- Button for ascending order -->
+    <Button
+        android:id="@+id/ascendingButton"
+        android:layout_width="wrap_content"
+        android:layout_height="match_parent"
+        android:layout_gravity=""
+        android:text="Ascending"
+        android:textColor="@android:color/white" />
+
+    <!-- Button for descending order -->
+
+</LinearLayout>
