+++ app/src/main/res/values/strings.xml
@@ -16,5 +16,6 @@
     <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="arrow_description">arrow to open details of an item</string>
     <string name="text_label">Item Deleted</string>
+    <string name="sort_method_name">Sort by Name</string>
 
 </resources>
